fix: ajuste no lancamento de excecoes de Procedimento

+++ microsservicos/conteudo/src/main/java/com/projetos/manutencao/conteudo/service/impl/ProcedimentoServiceImpl.java
@@ -1,5 +1,6 @@
 package com.projetos.manutencao.conteudo.service.impl;
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 import java.util.UUID;
 
@@ -25,14 +26,32 @@ public class ProcedimentoServiceImpl implements ProcedimentoService {
 
     @Override
     public Procedimento create(ProcedimentoDTO procedimentoDTO) {
+        if (procedimentoDTO == null) {
+            throw new IllegalArgumentException("Objeto ProcedimentoDTO não pode ser nulo.");
+        }
+
         Procedimento procedimento = modelMapper.map(procedimentoDTO, Procedimento.class);
-        procedimento.setId(UUID.randomUUID().toString());
+        
+        if (procedimento.getId() == null || procedimento.getId().isBlank()) {
+            procedimento.setId(UUID.randomUUID().toString());
+        }
+
         return repository.save(procedimento);
     }
 
     @Override
     public Optional<Procedimento> findById(String id) {
-        return repository.findById(id);
+        if (id == null || id.isBlank()) {
+            throw new IllegalArgumentException("ID não pode ser vazio.");
+        }
+
+        if (!repository.existsById(id)) {
+            throw new NoSuchElementException("Procedimento não encontrado para atualização: " + id);
+        }
+
+        Procedimento procedimento = repository.findById(id).get();
+
+        return Optional.of(procedimento);
     }
 
     @Override
@@ -42,13 +61,35 @@ public class ProcedimentoServiceImpl implements ProcedimentoService {
 
     @Override
     public Procedimento update(String id, ProcedimentoDTO procedimentoDTO) {
+
+        if (id == null || id.isBlank()) {
+            throw new IllegalArgumentException("ID não pode ser vazio.");
+        }
+
+        if (procedimentoDTO == null) {
+            throw new IllegalArgumentException("Objeto ProcedimentoDTO não pode ser nulo.");
+        }
+
+        if (!repository.existsById(id)) {
+            throw new NoSuchElementException("Procedimento não encontrado para atualização: " + id);
+        }
+
         Procedimento procedimento = modelMapper.map(procedimentoDTO, Procedimento.class);
+
         procedimento.setId(id);
         return repository.save(procedimento);
     }
 
     @Override
     public void deleteById(String id) {
+        if (id == null || id.isBlank()) {
+            throw new IllegalArgumentException("ID não pode ser vazio.");
+        }
+
+        if (!repository.existsById(id)) {
+            throw new NoSuchElementException("Procedimento não encontrado para atualização: " + id);
+        }
+
         repository.deleteById(id);
     }
 }
